vessel_management/requirements.txt
# Core Framework
Django==4.2.10
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7

# File Handling
Pillow==10.2.0
django-storages==1.14.2

# PDF Generation
reportlab>=3.6.12

# Email Services
sendgrid==6.11.0

# Development Tools
django-debug-toolbar==4.2.0
black==23.12.1
flake8==7.0.0
pytest==7.4.4
pytest-django==4.7.0

# Security
django-environ==0.11.2
django-axes==6.3.0

# Caching
django-redis==5.4.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Monitoring
sentry-sdk==1.39.1

# Testing
coverage==7.4.1
factory-boy==3.3.0



